small err fixed - some lines in scanner commented out

+++ src/jminusminus/Parser.java
@@ -304,7 +304,7 @@ public class Parser {
             return true;
         } else {
             scanner.recordPosition();
-            if (have(BOOLEAN) || have(CHAR) || have(INT) || have(DOUBLE)) {
+            if (have(BOOLEAN) || have(CHAR)  || have(DOUBLE) || have(INT)) {
                 if (have(LBRACK) && see(RBRACK)) {
                     scanner.returnToPosition();
                     return true;
